requirements.txt
django
djangorestframework
django-cors-headers
gunicorn
dj-database-url
whitenoise
psycopg2